Bugfix in CRS::alterCSLinearUnit for DerivedProjectedCRS
The implementation was calling alterUnit on the CS of the base
projected CRS, but this CS could be different than that of the
derived projected CRS being modified (e.g. the axes could be
swapped).

The fix consist on altering the unit of the derived projected
CRS, but only if the CS can be cast to Cartesian. It's doesn't
look like other CS types need to be considered.