Mention the proj_create function
Following the PROJ doc is like jumping through many hoops, or if you wish, many reprojections across documents are needed for something coherent to emerge. This addition adds a blurb mentioning the proj_create() function, which is really one of the two workhorses of this library, together with proj_create_crs_to_crs.