Fix silly documentation issue (quoted value of pi/2 triggers error)
The documentation included instances of "Example: +lat_1=1.5708r".
However if the poor user tries this setting, the result is "|lat_1|
should be < 90d".  This fix uses 1.570796r instead -- this is a
legitimate rounded value of pi/2 and, since it is less than pi/2, it
satisfies the requirement that it be less that 90d.  The examples with
the lon_0 are changed for consistency (even though the old value doesn't
trigger an error).