Implement the ellipsoidal gnomonic projection
This implements the ellipsoidal gnomonic projection given in
Karney (2013).  It is the limit of a doubly azimuthal projection, a
projection where the azimuths from 2 points are preserved, as the two
points merge into the center point.  Geodesics project to approximately
straight lines (these are exactly straight if the geodesic includes the
center point).

This change will potentially break some user code because the previous
implementation of the projection silently converted an ellipsoid into a
sphere with the same equatorial radius.  Now, it'll be treated as an
ellipsoid.  The previous (inconsistent!) behavior can be recovered by
setting the flattening to zero with "+f=0".

The boundary test in test_c_api.cpp needed to be updated because it
relied on this inconsistent treatment of the ellipsoid parameters.  The
change is explained in the file.

All the gnomonic tests in builtins.gie were supplemented by ellipsoidal
versions (with flattening = 1/200).

The documentation was updated to reflect this addition and the gnomonic
projection is listed as azimuthal (instead of, incorrectly, as
pseudocylindrical) and the defined area has been changed from "global"
to "within a quarter circumference of the center point".

Because the azimuthal equidistance projection uses the same geodesic
machinery as the new gnomonic projection, I'm simplified some of the
code in aeqd.cpp to parallel gnom.cpp; in particular, aeqd no longer
does the scaling/unscaling by P->a and the shifting/unshifting by
P->lam0.