replaced the Wagner 7 projection with a new version that has an inverse
The Wagner 7 projection had no inverse and thus was of no use for applications that needed an inverse projection, such as those which allow a user to click on a map and obtain the latitude and longitude. These new formulae are adapted from the D3 library, which has a permissive license, but more accurate constants, taken from a different source, are used. The new code has been tested and seems to work well.