createOperations(): support one of source/target to be a CoordinateMetadata
but not both.

Note also that the core of createOperations() does not take into account
the source/target coordinate epoch in its logic to infer pipelines. It
just keep those values in its metadata, so that PJ* objects constructed
from them can later override the PJ_COORD.xyzt.t component.

So basically this is currently useful only for transformations between a
dynamic CRS and a static CRS (or the reverse).