cs2cs / proj_create_crs_to_crs(): fix regression with geocentric CRS
Related to use case of https://github.com/OSGeo/gdal/issues/7753

When a transformation involved a datum shift with a grid and a
geocentric CRS as source/target, since PROJ 9.2.0 and the changes for
the only_best mode, even in just warning mode, if the grid of the "best"
transformation was missing, then the transformation would completely
fail, whereas with earlier versions it would fallback to a ballpark
transformation. So restore that mode, and actually enhance the operation
selection logic, by allowing to select operations based on their area of
validity even if the input coordinate are geocentric (in which case a
conversion of them to geographic is done).