Add stale GitHub action, as replacement for old stale GitHub integration
Previously we had the "stale bot" which posted reminders in inactive
issues and PR's. The stale bot used now removed functionality of GitHub
and it hasn't worked for about a year. This commit introduces the same
functionality based on GitHub Actions. The action setup is grabbed from
the QGIS project and adapted to the PROJ project.

With the stale action in place, PR's and issues are marked "stale" after
60 days of inactivity. If 7 more days passes without any activity the
issue or PR is closed.