CRS::stripVerticalComponent(): redirect it to demoteTo2D(std::string(), nullptr)
stripVerticalComponent() was implemented in the initial commit of the PROJ 6 / GDAL
Barn SRS effort, but was not used internally at that time. Just as a potential
useful method for future users

Later we needed promoteTo3D(), and soon after the reverse operation, but
we (likely) forgot about the existence of stripVerticalComponent(). Hence
a new demoteTo2D() method, which is more capable than the initial
stripVerticalComponent().