Fix clang-19 'warning: empty paragraph passed to '@throw' command [-Wdocumentation]'
clang-19 -Wdocumentation is super pedantic and requires an explanation
for why we throw exceptions. You're so demanding clang-19! We are kind
enough to mention we throw... Doxygen doesn't even require us to justify
the circumstances in which we have to do that. Oh well, let's please you
to get peace back...