createOperations(): use more appropriate operation when using a 'PROJ {grid_name}' geoid model, based on matching the vertical datum
Helps when transforming from 'NAD83(2011) / Puerto Rico and Virgin Is. + "VIVD09 height + PROJ us_noaa_g2012bp0.tif' to NAD83(2011).
Without that fix, PROJ would use the EPSG record for the NAD83(2011) <--> PRVD02 height (Puerto Rico) to get the extent of
the transformation, instead of the one for NAD83(2011) <--> VIVD09 height (Virgin Islands), causing the
resulting global transformation to not be usable in practice.