Use C++17 function ellint_1 in adams.cpp
Replace approximate series for F(x, 1/sqrt(2)) by a call to ellint_1.

The approximate series was only accurate to "better than 1e-7", which
meant that the errors in the projection were typically about 1 m.  The
adoption of ellint_1 (with its considerably better accuracy) then leads
to a torrent of errors in the tests.  These have been fixed by changing
the tolerances from 1 mm to 1 m.

The values in the tests should be replaced by more accurate ones, but
I'm reluctant merely to update the values with the ones the code now
produces given that the code depends on a parameter TOL = 1e-9.  Also, I
don't understand why K(1/sqrt(2)) = std::comp_ellint_1(RSQRT2) =
1.8540746773013719 is sometimes given to only 5 decimal places as
1.85407.