WKT/PROJJSON datum ensemble import: allow no members
With the future extension of the ETRS89 datum ensemble to all recently
created ETRS89-XX national datums, the number of members in the ensemble
will be huge and will make ugly WKT/PROJJSON, hence it is highly likely
that listing the members will be made optional. Be tolerant of that on
the parsing size, but still emit them on export for now.