Use Emscripten fetch in networkfilemanager (#4627)
The main purpose of this PR is an implementation in the networkfilemanager that substitutes cURL when running in wasm. For that purpose emscripten_fetch is used, as suggested in https://github.com/emscripten-core/emscripten/issues/3270#issuecomment-2260232768

The added implementation in `networkfilemanager.cpp` is a copy of the cURL implementation, trying to mimic it as much as possible. Emscripten does convert it into whatever code in wasm + js, using the fetch in the browser.

To enable this feature, PROJ must be compiled with the option `ENABLE_EMSCRIPTEN_FETCH`, that by default is OFF. It works only in `emscripten`.

As a needed tool for my tests I implemented a Docker image and a build script to produce a wasm compilation, and test it. It has been very useful, as it showed up some problems due to the synchronous fetch used in PROJ. Already having the build infrastructure, I added it as a github action. Anybody can take that to produce a wasm of PROJ.

[emscripten documentation](https://emscripten.org/docs/api_reference/fetch.html#synchronous-fetches) says clearly that to get synchronous fetch the `-pthread` flag is needed. It also needs that the javascript calls are done in a Web Worker, not in the main thread. I was not able to make `--proxy-to-worker` work, as there is no `main` to use it. The build script is then using the `-pthread` flag for all the dependencies that have to be compiled: `zlib`, `libtiff` and `libsqlite3`.
If a developer doesn't need the fetch functionality, it is much easier not linking with libtiff (and zlib).

If anybody with a better empscripten knowledge has a better solution, please let me know.

A basic example is running in https://jjimenezshaw.github.io/wasm-proj/example.html

For the usage in the browser, an special configuration for Cross-origin isolation is needed. As that is not included in GitHub Pages, I am using https://github.com/gzuidhof/coi-serviceworker in the example. Probably you need it too.